refactor and removal of unneeded calculators in bond-atom expansion.
