+++ rascaline/src/calculators/bondatom/mod.rs
@@ -1,10 +1,11 @@
-pub mod bondatom_neighbor_list;
 pub mod spherical_expansion_bondcentered;
-pub mod spherical_expansion_bondcentered_pair;
 
-pub use bondatom_neighbor_list::BANeighborList;
-pub use spherical_expansion_bondcentered::SphericalExpansionForBonds;
-pub use spherical_expansion_bondcentered_pair::{
-    SphericalExpansionForBondType,
-    SphericalExpansionForBondsParameters
+mod bond_atom_math;
+pub(crate) use bond_atom_math::canonical_vector_for_single_triplet;
+use bond_atom_math::{RawSphericalExpansion,RawSphericalExpansionParameters,ExpansionContribution};
+
+//pub use bondatom_neighbor_list::BANeighborList;
+pub use spherical_expansion_bondcentered::{
+    SphericalExpansionForBonds,
+    SphericalExpansionForBondsParameters,
 };
